Blog_Django_Project/pyproject.toml
[tool.poetry]
name = "blog-django-project"
version = "0.1.0"
description = ""
authors = ["Saytara2001 <lionheartsaytara@gmail.com>"]
readme = "README.md"
packages = [{include = "blog_django_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
